Simplify contrib handling
Since we don't accept arbitrary filenames, just the text files in the
share directory, there's no need to add a directory. This in turn
makes it unnecessary to qualify the files with a directory in the
example maps and the map generators, and that means we don't have to
pass the directory along to calls of generate_map.

We still need to pass it to the Mapper classes, however, since those
are the ones that do the actual parsing. They need to read that file
or request the URL.