Fix Alpine documentation, improved steepness
The Alpine documentation computed different defaults (notably for
peaks) than the actual algorithm. The map generating code now returns
the algorithm object if called in list contex and so the documentation
can get the actual values used for the documentation template.

Also changed some American spelling to British spelling (neighbors →
neighbours).

Made a change to how steepness is used: if steepness is higher (the
terrain therefore flatter), we search longer for candidates and do a
little random walk after a while to try and find neighbours. So now it
does make sense to use a steepness higher than 12.